+++ policyengine-client/src/countries/uk/uk.jsx
@@ -99,7 +99,19 @@ export class UK extends Country {
     },
     tv_licence_fee: {
       max: 250,
-    }
+    },
+    ofgem_price_cap_2022_q4: {
+      max: 10_000,
+    },
+    ofgem_price_cap_2023_q1: {
+      max: 10_000,
+    },
+    ofgem_price_cap_2023_q2: {
+      max: 10_000,
+    },
+    ofgem_price_cap_2023_q3: {
+      max: 10_000,
+    },
   };
   parameterHierarchy = {
     Simulation: {
@@ -107,7 +119,7 @@ export class UK extends Country {
       Geography: ["countrySpecific"],
     },
     "UK government": {
-      Tax: {
+      HMRC: {
         "Income Tax": {
           "Labour income": [
             "basic_rate",
@@ -117,9 +129,11 @@ export class UK extends Country {
             "add_threshold",
             "extra_UK_band",
           ],
+          "Dividend income": [
+            "gov_hmrc_income_tax_rates_dividends",
+          ],
           "Scottish rates": [
             "scottish_starter_rate",
-            "scottish_starter_threshold",
             "scottish_basic_rate",
             "scottish_basic_threshold",
             "scottish_intermediate_rate",
@@ -130,16 +144,22 @@ export class UK extends Country {
             "scottish_add_threshold",
             "extra_scot_band",
           ],
-          Allowances: [
-            "personal_allowance",
-            "PA_reduction_threshold",
-            "PA_reduction_rate",
-            "marriage_allowance_cap",
-            "abolish_marriage_allowance_income_condition",
-            "dividend_allowance",
-            "property_allowance",
-            "trading_allowance",
-          ],
+          Allowances: {
+            "Personal allowance": [
+              "personal_allowance",
+              "PA_reduction_threshold",
+              "PA_reduction_rate",
+            ],
+            "Marriage allowance": [
+              "marriage_allowance_cap",
+              "abolish_marriage_allowance_income_condition",
+            ],
+            "Other allowances": [
+              "dividend_allowance",
+              "property_allowance",
+              "trading_allowance",
+            ],
+          },
           Structural: ["abolish_income_tax"],
         },
         "National Insurance": {
@@ -159,7 +179,16 @@ export class UK extends Country {
           "abolish_lbtt",
           "abolish_business_rates",
         ],
+        "Child Benefit": [
+          "abolish_CB",
+          "CB_eldest",
+          "CB_additional",
+          "CB_HITC_reduction_threshold",
+          "CB_HITC_reduction_rate",
+        ],
         "Fuel duties": ["fuel_duty_rate"],
+      },
+      DCMS: {
         "TV licence": {
           "Fee": ["tv_licence_fee"],
           "Discounts": [
@@ -170,14 +199,7 @@ export class UK extends Country {
           "Evasion": ["tv_licence_evasion_rate"],
         }
       },
-      Benefit: {
-        "Child Benefit": [
-          "abolish_CB",
-          "CB_eldest",
-          "CB_additional",
-          "CB_HITC_reduction_threshold",
-          "CB_HITC_reduction_rate",
-        ],
+      DWP: {
         "Legacy benefits": [
           "abolish_CTC",
           "abolish_WTC",
@@ -216,12 +238,20 @@ export class UK extends Country {
             "UC_reduction_rate",
           ],
         },
+      },
+      Treasury: {
         "Energy bills support": ["ebr_ct_rebate", "ebr_energy_bills_credit"],
         "Cost-of-living support payment": [
           "col_benefit_payment_amount",
           "col_pensioner_payment_amount",
           "col_disability_payment_amount",
         ],
+        "Energy price cap subsidy": [
+          "ofgem_price_cap_2022_q4",
+          "ofgem_price_cap_2023_q1",
+          "ofgem_price_cap_2023_q2",
+          "ofgem_price_cap_2023_q3",
+        ]
       },
     },
     "Third party": {
@@ -264,7 +294,7 @@ export class UK extends Country {
     },
   };
   defaultOpenParameterGroups = [];
-  defaultSelectedParameterGroup = "/UK government/Tax/Income Tax/Labour income";
+  defaultSelectedParameterGroup = "/UK government/HMRC/Income Tax/Labour income";
   organisations = {
     "UBI Center": {
       logo: UBICenterLogo,
@@ -321,21 +351,21 @@ export class UK extends Country {
   situation = {
     people: {
       You: {
-        age: { 2022: 25 },
+        age: { [this.year]: 25 },
       },
     },
     benunits: {
       "Your family": {
         adults: ["You"],
         children: [],
-        claims_all_entitled_benefits: { 2022: true },
+        claims_all_entitled_benefits: { [this.year]: true },
       },
     },
     households: {
       "Your household": {
         adults: ["You"],
         children: [],
-        household_owns_tv: { 2022: true },
+        household_owns_tv: { [this.year]: true },
       },
     },
     states: {
@@ -528,10 +558,10 @@ export class UK extends Country {
 
   addPartner(situation) {
     situation.people["Your spouse"] = {
-      age: { 2022: 25 },
+      age: { [this.year]: 25 },
     };
     situation.benunits["Your family"].adults.push("Your spouse");
-    situation.benunits["Your family"]["is_married"]["2022"] = true;
+    situation.benunits["Your family"]["is_married"][this.year] = true;
     situation.households["Your household"].adults.push("Your spouse");
     return this.validateSituation(situation).situation;
   }
@@ -540,7 +570,7 @@ export class UK extends Country {
     const childName =
       childNamer[situation.benunits["Your family"].children.length + 1];
     situation.people[childName] = {
-      age: { 2022: 10 },
+      age: { [this.year]: 10 },
     };
     situation.benunits["Your family"].children.push(childName);
     situation.households["Your household"].children.push(childName);
@@ -563,7 +593,7 @@ export class UK extends Country {
       situation.households["Your household"].children.pop(name);
     }
     if (name === "Your spouse") {
-      situation.benunits["Your family"]["is_married"]["2022"] = false;
+      situation.benunits["Your family"]["is_married"][this.year] = false;
     }
     delete situation.people[name];
     return this.validateSituation(situation).situation;
@@ -624,5 +654,11 @@ export class UK extends Country {
     "extra_scot_rate",
     "extra_scot_threshold",
     "country_specific",
+    "dividend_basic_rate",
+    "dividend_basic_threshold",
+    "dividend_higher_rate",
+    "dividend_higher_threshold",
+    "dividend_additional_rate",
+    "dividend_additional_threshold",
   ];
 }