+++ policyengine-client/src/policyengine/pages/household/accountingTable.jsx
@@ -133,11 +133,11 @@ function getValues(variable, country) {
     const entity = country.entities[country.variables[variable].entity];
     const entities = Object.keys(baseline[entity.plural]);
     const baselineValue = entities.length > 1 ?
-        entities.map(name => baseline[entity.plural][name][variable]["2022"]).reduce((a, b) => a + b, 0) :
-        baseline[entity.plural][entities[0]][variable]["2022"];
+        entities.map(name => baseline[entity.plural][name][variable][country.year]).reduce((a, b) => a + b, 0) :
+        baseline[entity.plural][entities[0]][variable][country.year];
     const reformValue = entities.length > 1 ?
-        entities.map(name => reform[entity.plural][name][variable]["2022"]).reduce((a, b) => a + b, 0) :
-        reform[entity.plural][entities[0]][variable]["2022"];
+        entities.map(name => reform[entity.plural][name][variable][country.year]).reduce((a, b) => a + b, 0) :
+        reform[entity.plural][entities[0]][variable][country.year];
     return { baselineValue: baselineValue, reformValue: reformValue };
 }