+++ policyengine-client/src/policyengine/pages/household/menu.jsx
@@ -31,10 +31,9 @@ export default function Menu(props) {
 			mode="inline"
 			defaultOpenKeys={country.defaultOpenVariableGroups}
 			defaultSelectedKeys={[country.defaultSelectedVariableGroup]}
+			style={{fontSize: 18, paddingBottom: 20}}
 		>
-            <Divider orientation="left" onClick={() => props.selectVariableGroup(country.defaultSelectedVariableGroup)}>About your household</Divider>
 			{addMenuEntry(country.inputVariableHierarchy, "")}
-
 		</AntMenu>
-	);
+	)
 }