+++ policyengine/country/uk/default_reform.py
@@ -24,7 +24,7 @@ baseline_parameters = CountryTaxBenefitSystem().parameters
 
 
 def add_extra_band(parameters: ParameterNode) -> ParameterNode:
-    rates = parameters.tax.income_tax.rates
+    rates = parameters.gov.hmrc.income_tax.rates
     uk_rates: ParameterScale = rates.uk
     extra_uk_bracket = ParameterScaleBracket(
         data={
@@ -310,6 +310,54 @@ def create_default_reform() -> ReformType:
                 return where(is_sp_age, baseline_pa, default_pa)
             return default_pa
 
+    class energy_price_cap_subsidy(
+        baseline_variables["energy_price_cap_subsidy"]
+    ):
+        def formula(household, period, parameters):
+            # PolicyEngine simulates each month, rather than each quarter in OpenFisca UK. This
+            # is because PolicyEngine simulations are of "the year from now", rather than a specific
+            # year.
+            energy_consumption = household(
+                "domestic_energy_consumption", period
+            )
+            country = household("country", period)
+            outside_ni = country != country.possible_values.NORTHERN_IRELAND
+            energy_consumption = max_(0, energy_consumption)
+            # For each of the four quarters in the next year, calculate the
+            # relative change to the price cap against the baseline price cap,
+            # and multiply by quarterly energy consumption.
+            total_subsidy = 0
+            # Manually use the next four quarters (from now) in place of the first four quarters of the current year.
+            now = datetime.now()
+            m1_baseline_price_cap = parameters(
+                period
+            ).baseline.gov.ofgem.price_cap._children[f"{now.year}_q1"]
+            for month_from_now in range(1, 13):
+                into_next_year = now.month + month_from_now > 12
+                month = (now.month + month_from_now) % 12
+                year = now.year + into_next_year * 1
+                quarter = month // 3 + 1
+                current_quarter = f"{year}_q{quarter}"
+                price_cap = parameters(period).gov.ofgem.price_cap._children[
+                    current_quarter
+                ]
+                baseline_price_cap = parameters(
+                    period
+                ).baseline.gov.ofgem.price_cap._children[current_quarter]
+                relative_change_in_cap = (
+                    price_cap - baseline_price_cap
+                ) / baseline_price_cap
+                relative_change_in_energy_consumption = (
+                    baseline_price_cap / m1_baseline_price_cap
+                )
+                total_subsidy += (
+                    -relative_change_in_cap
+                    * relative_change_in_energy_consumption
+                    * energy_consumption
+                    / 12
+                )
+            return total_subsidy * outside_ni
+
     class default_reform(Reform):
         def apply(self):
             self.update_variable(LVT)
@@ -329,6 +377,6 @@ def create_default_reform() -> ReformType:
             add_parameter_file(
                 Path(__file__).parent / "additional_parameters.yaml"
             ).apply(self)
-            use_current_parameters().apply(self)
+            self.update_variable(energy_price_cap_subsidy)
 
-    return default_reform
+    return default_reform, use_current_parameters()